requirements.txt
mkdocs==1.0.4
pymdown-extensions==6.0
mkdocs-material>=2.8.0
google-api-python-client==1.7.8
oauth2client==4.1.3
progressbar2==3.39.3
exabyte_json_include==2.0.0
git+git://github.com/Exabyte-io/markdown-include.git#egg=markdown_include
git+git://github.com/tokland/youtube-upload.git
